+++ src/lib/output/themes/default/assets/typedoc/Navigation.ts
@@ -34,7 +34,8 @@ async function buildNav() {
         .pipeThrough(new DecompressionStream("gzip"));
     const nav: NavigationElement[] = await new Response(json).json();
 
-    BASE_URL = container.dataset.base + "/";
+    BASE_URL = container.dataset.base!;
+    if (!BASE_URL.endsWith("/")) BASE_URL += "/";
     container.innerHTML = "";
     for (const el of nav) {
         buildNavElement(el, container, []);